New contexts added

+++ src/main/java/com/sda/spring/demo/model/Author.java
@@ -1,14 +1,9 @@
 package com.sda.spring.demo.model;
 
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @Setter
@@ -18,7 +13,7 @@ public class Author {
     @Getter
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     @Getter
     private String name;
     @Getter
