+++ src/main/java/com/sda/spring/demo/model/Category.java
@@ -1,25 +1,22 @@
 package com.sda.spring.demo.model;
 
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @NoArgsConstructor
 @Setter
 @Entity
 public class Category {
+
     @Getter
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "category_ID")
-    private int id;
+    private Long id;
+
     @Getter
     private String name;
